Bug 1315784 - Reorganize L10N Screenshot Tests
